Add files via upload
This project is about visualizing US presidential election data, with interactive control of a timeline, and how the states are colored according to their votes. There are three main modes:

'RVD' (republican-vs-democratic): the scatterplot shows Republican votes increasing left to right (specifically, the X location is determined by a warping function on the number of Republican votes), democratic votes increasing bottom to top (as determined by a warping function on the number of democratic votes), and the colormap is simply either red or blue according to whichever one is larger

'PUR' (shades of purple): the scatterplot is the same as in RVD.  The colormap goes between blue and red through some purple

'LVA' (lean-vs-amount): X direction of the scatterplot shows political leaning, and Y-direction shows the overall voting amount.